Add top-level grid
Has sidebar, header, and main sections.

I chose to have the DOM order go header > main > sidebar because to
me that sounds like it would be the most reasonable for tab
navigation. I could be mistaken about that, but I think it's a fair
first guess.

When describing the grid areas for each section, I tried to define
them in such a way that it would still behave well if there was only
one column. I did this by telling the header and main to use the
"last column" and telling the sidebar to fill in the remaining
space on the left, despite being last in the DOM order.

It isn't currently implemented, but I think this will make things
more gracefully able to transition to a responsive, one-column
layout when the time comes.